variantMinLength and variantMaxLength
This PR introduces additional, optional variant query parameters which have been identified by the Beacon Variant Scout Team to enable new types of variant queries. Examples would be here e.g. "a deletion of at least 10bp in the 5` region of a gene", where the genomic region can be scoped through a range or bracket query, but the size cannot.